// api/src/user/user-response.dto.ts
export class UserResponseDTO {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  avatar: string;
  isEmailVerified: boolean;
  bio: string;
  notifications: number;
  age: string;
  gender: string;
  major: string;
  gradYear: string;
  stayLength: string;
  budget: string;
  idealDistance: string;
  petPreference: string;
  cleanliness: string;
  smoker: string;
  socialPreference: string;
  peakProductivity: string;
}